+++ src/jsMain/kotlin/com/wadii/pages/admin/service/ServicesEvent.kt
@@ -9,7 +9,7 @@ sealed class ServicesEvent {
     object CancelEdit : ServicesEvent()
     data class SetEditName(val name: String) : ServicesEvent()
     data class SaveEdit(val service: Service) : ServicesEvent()
-    data class Delete(val serviceId: Int) : ServicesEvent()
+    data class Delete(val serviceId: Long) : ServicesEvent()
     data class SetNewName(val name: String) : ServicesEvent()
     object Add : ServicesEvent()
 }
